Crate nix_bindings_store_sys

Raw bindings to Nix C API

This crate contains automatically generated bindings from the Nix C headers. The bindings are generated by bindgen and include C-style naming conventions and documentation comments that don’t always conform to Rust standards.

Normally you don’t have to use this crate directly. Instead use nix-store.

Structs

Store

StorePath

__fsid_t

derivation

max_align_t

store_path_hash_part

@brief A store path hash

Constants

INT8_MAX

INT8_MIN

INT8_WIDTH

INT16_MAX

INT16_MIN

INT16_WIDTH

INT32_MAX

INT32_MIN

INT32_WIDTH

INT64_WIDTH

INTMAX_WIDTH

INTPTR_MAX

INTPTR_MIN

INTPTR_WIDTH

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST8_WIDTH

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST16_WIDTH

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST32_WIDTH

INT_FAST64_WIDTH

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST8_WIDTH

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST16_WIDTH

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST32_WIDTH

INT_LEAST64_WIDTH

PTRDIFF_MAX

PTRDIFF_MIN

PTRDIFF_WIDTH

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIG_ATOMIC_WIDTH

SIZE_MAX

SIZE_WIDTH

UINT8_MAX

UINT8_WIDTH

UINT16_MAX

UINT16_WIDTH

UINT32_MAX

UINT32_WIDTH

UINT64_WIDTH

UINTMAX_WIDTH

UINTPTR_MAX

UINTPTR_WIDTH

UINT_FAST8_MAX

UINT_FAST8_WIDTH

UINT_FAST16_MAX

UINT_FAST16_WIDTH

UINT_FAST32_MAX

UINT_FAST32_WIDTH

UINT_FAST64_WIDTH

UINT_LEAST8_MAX

UINT_LEAST8_WIDTH

UINT_LEAST16_MAX

UINT_LEAST16_WIDTH

UINT_LEAST32_MAX

UINT_LEAST32_WIDTH

UINT_LEAST64_WIDTH

WCHAR_WIDTH

WINT_MAX

WINT_MIN

WINT_WIDTH

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_DYNAMIC_STACK_SIZE_SOURCE

_FEATURES_H

_ISOC2Y_SOURCE

_ISOC11_SOURCE

_ISOC23_SOURCE

_ISOC95_SOURCE

_ISOC99_SOURCE

_LARGEFILE64_SOURCE

_LARGEFILE_SOURCE

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

_XOPEN_SOURCE

_XOPEN_SOURCE_EXTENDED

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C23_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C23

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C23

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2Y

__GLIBC_USE_ISOC23

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_DYNAMIC_STACK_SIZE

__USE_FORTIFY_LEVEL

__USE_GNU

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_ISOCXX11

__USE_LARGEFILE

__USE_LARGEFILE64

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_TIME_BITS64

__USE_UNIX98

__USE_XOPEN

__USE_XOPEN2K

__USE_XOPEN2K8

__USE_XOPEN2K8XSI

__USE_XOPEN2KXSI

__USE_XOPEN_EXTENDED

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

Functions

add_derivation⚠

@brief Add the given nix_derivation to the given store

derivation_clone⚠

@brief Copy a nix_derivation

derivation_free⚠

@brief Deallocate a nix_derivation

derivation_from_json⚠

@brief Create a nix_derivation from a JSON representation of that derivation.

derivation_to_json⚠

@brief Gets the derivation as a JSON string

libstore_init⚠

@brief Initializes the Nix store library

libstore_init_no_load_config⚠

@brief Like nix_libstore_init, but does not load the Nix configuration.

store_copy_closure⚠

@brief Copy the closure of path from srcStore to dstStore.

store_create_from_parts⚠

@brief Create a StorePath from its constituent parts (hash and name)

store_drv_from_store_path⚠

@brief Returns the derivation associated with the store path

store_free⚠

@brief Deallocate a nix store and free any resources if not also held by other Store instances.

store_get_fs_closure⚠

@brief Gets the closure of a specific store path

store_get_storedir⚠

@brief get the storeDir of a Nix store, typically "/nix/store" @param[out] context Optional, stores error information @param[in] store nix store reference @param[in] callback Called with the URI. @param[in] user_data optional, arbitrary data, passed to the callback when it’s called. @see nix_get_string_callback @return error code, NIX_OK on success.

store_get_uri⚠

@brief get the URI of a nix store @param[out] context Optional, stores error information @param[in] store nix store reference @param[in] callback Called with the URI. @param[in] user_data optional, arbitrary data, passed to the callback when it’s called. @see nix_get_string_callback @return error code, NIX_OK on success.

store_get_version⚠

@brief get the version of a nix store.

store_is_valid_path⚠

@brief Check if a StorePath is valid (i.e. that corresponding store object and its closure of references exists in the store) @param[out] context Optional, stores error information @param[in] store Nix Store reference @param[in] path Path to check @return true or false, error info in context

store_open⚠

@brief Open a nix store.

store_parse_path⚠

@brief Parse a Nix store path that includes the store dir into a StorePath

store_path_clone⚠

@brief Copy a StorePath

store_path_free⚠

@brief Deallocate a StorePath

store_path_hash⚠

@brief Get the path hash (e.g. “” in /nix/store/-)

store_path_name⚠

@brief Get the path name (e.g. “” in /nix/store/-)

store_query_path_from_hash_part⚠

@brief Query the full store path given the hash part of a valid store path, or empty if no matching path is found.

store_real_path⚠

@brief Get the physical location of a store path

store_realise⚠

@brief Realise a Nix store path

Type Aliases

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t